KVStore: Fix unreleased Region instance (#9763)
close pingcap/tiflash#9722
