+++ dbms/src/Storages/KVStore/MultiRaft/RegionState.h
@@ -66,6 +66,5 @@ private:
 HandleRange<HandleID> getHandleRangeByTable(
     const std::pair<DecodedTiKVKeyPtr, DecodedTiKVKeyPtr> & rawKeys,
     TableID table_id);
-bool computeMappedTableID(const DecodedTiKVKey & key, TableID & table_id);
 
 } // namespace DB
